// src/Components/Model/eFilm.tsx
export default interface eFilm {
  id: string;
  name: string;
  subName: string;
  description: string;
  trailer: string;
  startdate: string;
  imagePortrait: string;
  imageLandscapeMobile: string;
  duration: string;
  point: number;
  age: string;
  imageLandscape: string;
  shortDescription: string
}